feat: Adicionar gerenciamento de lojas com rotas e controladores
Implementa a funcionalidade de gerenciamento de lojas, incluindo a criação, busca por ID e slug, e rotas correspondentes. Adiciona validação de entrada e tratamento de erros, melhorando a robustez da API. Integra as novas rotas no arquivo index.ts e atualiza o repositório de lojas para suportar as operações necessárias, facilitando a interação com a base de dados.

+++ src/use-cases/users/delete-user.ts
@@ -1,4 +1,5 @@
 import type { UsersRepository } from "~/repositories/users-repository";
+import { UserNotFoundError } from "../@errors/users/user-not-found-error";
 
 interface DeleteUserUseCaseRequest {
 	id: string;
@@ -8,6 +9,12 @@ export class DeleteUserUseCase {
 	constructor(private readonly usersRepository: UsersRepository) {}
 
 	async execute({ id }: DeleteUserUseCaseRequest): Promise<void> {
+		const user = await this.usersRepository.findById({ id });
+
+		if (!user) {
+			throw new UserNotFoundError();
+		}
+
 		await this.usersRepository.delete({ id });
 	}
 }
